Full support for PyQt4, PyQt5, and Pyside at runtime
This abstracts the differences between the PyQt4 and PyQt5 module differences on
a case-by-case basis. It's fragile, but works.  To do this, UI code is no longer
generated and is instead loaded at runtime.  This introduces another
incompatability with PySide because it does not have a uic module.  To address
this a pyside_dynamic abstraction is introduced, all wrapped in
load_ui_widget().